Replaces integer(4) by integer

+++ examples/cairo-tests.f90
@@ -179,7 +179,7 @@ contains
     use, intrinsic :: iso_fortran_env, only: wp=>real64, int8
 
     type(c_ptr) :: my_drawing_area
-    integer(4) :: i, j, k, p, itermax
+    integer    :: i, j, k, p, itermax
     real(8)    :: x, y, xmin, xmax, ymin, ymax ! coordinates in the complex plane
     complex(8) :: c, z
     real(8)    :: scx, scy             ! scales
